feat: enhance CourseController and CourseService with improved course download functionality, including detailed storage validation and additional query parameters for estimating course size, while introducing new services for category, comment, and additional resource management
