+++ src/(resources)/course/services/category.service.ts
@@ -1,0 +1,171 @@
+import {
+  BadRequestException,
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { Category } from '../entities/category.entity';
+import { Course } from '../entities/course.entity';
+import {
+  CreateCategoryDto,
+  UpdateCategoryDto,
+  CategoryResponseDto,
+} from '../dto/category.dto';
+
+@Injectable()
+export class CategoryService {
+  constructor(
+    @InjectRepository(Category)
+    private readonly categoryRepository: Repository<Category>,
+    @InjectRepository(Course)
+    private readonly courseRepository: Repository<Course>,
+  ) {}
+
+  async createCategory(
+    createCategoryDto: CreateCategoryDto,
+  ): Promise<Category> {
+    try {
+      // Check if category with the same name already exists
+      const existingCategory = await this.categoryRepository.findOne({
+        where: { name: createCategoryDto.name },
+      });
+
+      if (existingCategory) {
+        throw new ConflictException(
+          `Category with name "${createCategoryDto.name}" already exists`,
+        );
+      }
+
+      const category = this.categoryRepository.create(createCategoryDto);
+      return await this.categoryRepository.save(category);
+    } catch (error) {
+      if (error instanceof ConflictException) {
+        throw error;
+      }
+      throw new BadRequestException(
+        'Failed to create category: ' + error.message,
+      );
+    }
+  }
+
+  async findAllCategories(): Promise<CategoryResponseDto[]> {
+    const categories = await this.categoryRepository.find();
+
+    if (!categories) {
+      throw new NotFoundException('No categories found');
+    }
+
+    // Get course counts for each category
+    const categoriesWithCounts = await Promise.all(
+      categories.map(async (category) => {
+        const courseCount = await this.courseRepository.count({
+          where: { categoryId: category.id },
+        });
+
+        return {
+          id: category.id,
+          name: category.name,
+          description: category.description,
+          imageUrl: category.imageUrl,
+          courseCount,
+        };
+      }),
+    );
+
+    return categoriesWithCounts;
+  }
+
+  async findOneCategory(id: string): Promise<CategoryResponseDto> {
+    const category = await this.categoryRepository.findOne({
+      where: { id },
+      relations: {
+        courses: true,
+      },
+    });
+
+    if (!category) {
+      throw new NotFoundException(`Category with ID ${id} not found`);
+    }
+
+    const courseCount = await this.courseRepository.count({
+      where: { categoryId: id },
+    });
+
+    return {
+      id: category.id,
+      name: category.name,
+      description: category.description,
+      imageUrl: category.imageUrl,
+      courseCount,
+    };
+  }
+
+  async updateCategory(
+    id: string,
+    updateCategoryDto: UpdateCategoryDto,
+  ): Promise<Category> {
+    const category = await this.categoryRepository.findOne({
+      where: { id },
+    });
+
+    if (!category) {
+      throw new NotFoundException(`Category with ID ${id} not found`);
+    }
+
+    // If name is being updated, check for conflicts
+    if (updateCategoryDto.name && updateCategoryDto.name !== category.name) {
+      const existingCategory = await this.categoryRepository.findOne({
+        where: { name: updateCategoryDto.name },
+      });
+
+      if (existingCategory) {
+        throw new ConflictException(
+          `Category with name "${updateCategoryDto.name}" already exists`,
+        );
+      }
+    }
+
+    Object.assign(category, updateCategoryDto);
+    return await this.categoryRepository.save(category);
+  }
+
+  async removeCategory(id: string): Promise<void> {
+    const category = await this.categoryRepository.findOne({
+      where: { id },
+    });
+
+    if (!category) {
+      throw new NotFoundException(`Category with ID ${id} not found`);
+    }
+
+    // Check if there are courses using this category
+    const courseCount = await this.courseRepository.count({
+      where: { categoryId: id },
+    });
+
+    if (courseCount > 0) {
+      throw new BadRequestException(
+        `Cannot delete category that contains ${courseCount} courses. Remove or reassign courses first.`,
+      );
+    }
+
+    await this.categoryRepository.remove(category);
+  }
+
+  async getCoursesByCategory(categoryId: string): Promise<Course[]> {
+    const category = await this.categoryRepository.findOne({
+      where: { id: categoryId },
+    });
+
+    if (!category) {
+      throw new NotFoundException(`Category with ID ${categoryId} not found`);
+    }
+
+    return await this.courseRepository.find({
+      where: { categoryId },
+      relations: ['user'],
+    });
+  }
+}
