+++ src/(resources)/course/services/index.ts
@@ -1,0 +1,7 @@
+export * from './category.service';
+export * from './downloadable-resource.service';
+export * from './quiz.service';
+export * from './comment.service';
+export * from './additional-resource.service';
+export * from './course-progress.service';
+export * from './storage-calculator.service';
